refactor : enum 클래스 패키지 변경
enum 클래스 패키지 변경
- file 패키지 <- FIlePath.java
- message 패키지 <- OuputMessage.java, ErrorMessage.java
- promotion 패키지 <- PromotionType.java
