+++ src/main/java/store/constant/message/OutputMessage.java
@@ -1,4 +1,4 @@
-package store.constant;
+package store.constant.message;
 
 public enum OutputMessage {
 	PROMPT_CURRENT_STOCK("안녕하세요. W편의점입니다.\n현재 보유하고 있는 상품입니다."),
